requirements.txt
fastapi==0.115.2
uvicorn==0.31.1
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.9.2
typing-extensions==4.12.2
starlette==0.40.0
aiofiles==23.2.1
email-validator==2.1.0  
pydantic-core==2.23.4
aiosmtplib==2.0.1
python-dotenv==1.0.0